自学训练营学习7群

+++ exercises/1901050079/1001S02E03_calculator.py
@@ -1,9 +1,9 @@
 print('输入数字开始进行计算')
-a=int(input(''))
+a=float(input(''))
 print('请选择：\n+\n-\n*\n/')
 c=input('')
 print('请继续输入数字')
-b=int(input(''))
+b=float(input(''))
 if c=='+':
     print(a+b)
 elif c=='-':
